Split image generation completion persistence
