+++ frontend/src/server/images/execute-image-generation.ts
@@ -19,11 +19,8 @@ import {
   createSignedDownloadUrl,
   extractStorageKeyFromUrl,
   isStorageConfigured,
-  recordUserAsset,
 } from '@/server/storage';
 import { createImageThumbnailBatch } from '@/server/image-thumbnails';
-import { buildStoredImageRenderEntries, resolveHeroThumbFromRenders } from '@/lib/image-renders';
-import { ensureReusableAsset, upsertLegacyJobOutputs } from '@/server/media-library';
 import {
   formatSupportedImageFormatsLabel,
   getSupportedImageFormats,
@@ -54,7 +51,8 @@ import {
   type StoredAssetInfo,
 } from './image-reference-normalization';
 import { ImageGenerationExecutionError } from './image-generation-error';
-import { createAtomicInitialImageJob, PLACEHOLDER_THUMB } from './image-initial-job';
+import { createAtomicInitialImageJob } from './image-initial-job';
+import { persistCompletedImageGeneration } from './image-generation-completion';
 import {
   buildCharacterReferencePrompt,
   extractImages,
@@ -693,150 +691,41 @@ export async function executeImageGeneration({
       ...image,
       thumbUrl: thumbUrls[index] ?? null,
     }));
-    const storedRenderEntries = buildStoredImageRenderEntries(normalizedImagesWithThumbs);
-    const hero = normalizedImagesWithThumbs[0]?.url ?? null;
-    const heroThumb = resolveHeroThumbFromRenders(normalizedImagesWithThumbs) ?? hero;
     const description =
       (result.data && typeof result.data === 'object' && (result.data as { description?: string }).description) || null;
-    const renderIdsJson = JSON.stringify(storedRenderEntries);
     const providerRequestId = providerJobId ?? parseRequestId(result.data) ?? result.requestId ?? null;
     providerJobId = providerRequestId ?? undefined;
 
-    if (jobSurface === 'character') {
-      await Promise.all(
-        normalizedImagesWithThumbs.map(async (image, index) => {
-          try {
-            await recordUserAsset({
-              assetId: `${jobId}:character:${index + 1}`,
-              userId,
-              url: image.url,
-              mime: image.mimeType ?? 'image/png',
-              width: image.width ?? null,
-              height: image.height ?? null,
-              size: null,
-              source: 'character',
-              metadata: {
-                jobId,
-                thumbUrl: image.thumbUrl ?? null,
-                engineId: engine.id,
-                engineLabel: engine.label,
-                surface: jobSurface,
-                billingProductKey,
-                resultIndex: index,
-              },
-            });
-          } catch (assetError) {
-            console.warn('[images] failed to persist character asset', { jobId, index }, assetError);
-          }
-        })
-      );
-    }
-
-    await query(
-      `UPDATE app_jobs
-       SET thumb_url = $2,
-           video_url = $3,
-           status = 'completed',
-           progress = 100,
-           provider_job_id = COALESCE($4, provider_job_id),
-           final_price_cents = $5,
-           pricing_snapshot = $6::jsonb,
-           cost_breakdown_usd = $7::jsonb,
-           currency = $8,
-           vendor_account_id = $9,
-           payment_status = 'paid_wallet',
-           visibility = $10,
-           indexable = $11,
-           message = $12,
-           render_ids = $13::jsonb,
-           hero_render_id = $14,
-           provisional = FALSE,
-           updated_at = NOW()
-       WHERE job_id = $1`,
-      [
-        jobId,
-        heroThumb ?? PLACEHOLDER_THUMB,
-        null,
-        providerJobId ?? null,
-        pricing.totalCents,
-        pricingSnapshotJson,
-        costBreakdownJson,
-        pricing.currency,
-        vendorAccountId,
-        visibility,
-        indexable,
-        description,
-        renderIdsJson,
-        hero,
-      ]
-    );
-
-    await upsertLegacyJobOutputs({
-      job_id: jobId,
-      user_id: userId,
-      surface: jobSurface,
-      video_url: null,
-      audio_url: null,
-      thumb_url: heroThumb ?? PLACEHOLDER_THUMB,
-      preview_frame: heroThumb ?? PLACEHOLDER_THUMB,
-      render_ids: storedRenderEntries,
-      duration_sec: 1,
-      status: 'completed',
-    }).catch((outputError) => {
-      console.warn('[images] failed to persist job outputs', { jobId }, outputError);
+    const { heroThumb } = await persistCompletedImageGeneration({
+      billingProductKey,
+      characterReferenceCount: characterReferences.length,
+      costBreakdownJson,
+      description,
+      enableWebSearch,
+      engineId: engine.id,
+      engineLabel: engine.label,
+      images: normalizedImagesWithThumbs,
+      indexable,
+      jobId,
+      jobSurface,
+      limitGenerations,
+      maskUrl,
+      mode,
+      normalizedSeed,
+      numImages,
+      outputFormat,
+      pricing,
+      pricingSnapshotJson,
+      providerJobId: providerJobId ?? null,
+      providerMode,
+      quality,
+      resolvedAspectRatio,
+      resolution,
+      thinkingLevel,
+      userId,
+      vendorAccountId,
+      visibility,
     });
-
-    if (jobSurface === 'character') {
-      await Promise.allSettled(
-        normalizedImagesWithThumbs.map((image, index) =>
-          ensureReusableAsset({
-            userId,
-            url: image.url,
-            kind: 'image',
-            source: 'character',
-            sourceJobId: jobId,
-            sourceOutputId: `${jobId}:image:${index}`,
-            mimeType: image.mimeType ?? 'image/png',
-            width: image.width ?? null,
-            height: image.height ?? null,
-            thumbUrl: image.thumbUrl ?? null,
-          })
-        )
-      );
-    }
-
-    try {
-      await query(
-        `INSERT INTO fal_queue_log (job_id, provider, provider_job_id, engine_id, status, payload)
-         VALUES ($1,$2,$3,$4,$5,$6::jsonb)`,
-        [
-          jobId,
-          providerMode,
-          providerJobId ?? null,
-          engine.id,
-          'completed',
-          JSON.stringify({
-            request: {
-              mode,
-              numImages,
-              resolution,
-              ...(characterReferences.length ? { characterReferenceCount: characterReferences.length } : {}),
-              ...(resolvedAspectRatio ? { aspect_ratio: resolvedAspectRatio } : {}),
-              ...(normalizedSeed != null ? { seed: normalizedSeed } : {}),
-              ...(outputFormat ? { output_format: outputFormat } : {}),
-              ...(quality ? { quality } : {}),
-              ...(maskUrl ? { mask_url: maskUrl } : {}),
-              ...(enableWebSearch ? { enable_web_search: true } : {}),
-              ...(thinkingLevel ? { thinking_level: thinkingLevel } : {}),
-              ...(limitGenerations ? { limit_generations: true } : {}),
-            },
-            pricing: { totalCents: pricing.totalCents, currency: pricing.currency },
-          }),
-        ]
-      );
-    } catch (error) {
-      console.warn('[images] failed to record fal queue log', error);
-    }
 
     return {
       ok: true,
